* start menu now is a little nicer and contains the current window selection as well * csv export fixed * csv export can now also export current selection * now current recording shows different background color too * TimeTracker now also shows status menu when timer is running (stop/edit) * now tracker does not update whole list of work items every second the timer shoots * User can now filter for weekday

+++ src/objc/TMetaProject.m
@@ -64,7 +64,7 @@
 	
 	while (aProject = [enumerator nextObject])
 	{
-		unsigned result = [[aProject tasks] indexOfObject:task];
+		NSUInteger result = [[aProject tasks] indexOfObject:task];
 		if (result != NSNotFound) {
 			*taskIndex = result;
 			return aProject;
